split crud and models into forecasts and observations
