+++ api/Makefile
@@ -112,6 +112,13 @@ run-bender-raw:
 run-marvin-raw:
 	python -m app.fireweather_bot.hourly_actuals
 
+run-all-bots-raw:
+	python -m app.fireweather_bot.noon_forecasts
+	python -m app.fireweather_bot.hourly_actuals
+	POSTGRES_HOST=localhost python -m app.models.env_canada GDPS
+	POSTGRES_HOST=localhost python -m app.models.env_canada HRDPS
+	POSTGRES_HOST=localhost python -m app.models.env_canada RDPS
+
 notebook:
 	# Run jupyter notebooks.
 	POSTGRES_HOST=localhost PYTHONPATH=$(shell pwd) JUPYTER_PATH=$(shell pwd) poetry run jupyter notebook --ip 0.0.0.0
