+++ api/app/models/machine_learning.py
@@ -11,7 +11,8 @@ import numpy as np
 from sqlalchemy.orm import Session
 from app.models import MODEL_VALUE_KEYS, construct_interpolated_noon_prediction
 from app.db.models import (
-    PredictionModel, PredictionModelGridSubset, ModelRunGridSubsetPrediction, HourlyActual)
+    PredictionModel, PredictionModelGridSubset, ModelRunGridSubsetPrediction)
+from app.db.models.observations import HourlyActual
 from app.db.crud.observations import get_actuals_left_outer_join_with_predictions
