+++ api/app/tests/models/crud.py
@@ -1,7 +1,8 @@
 """ Some crud responses used to mock our calls to app.db.crud
 """
 from datetime import datetime
-from app.db.models import HourlyActual, ModelRunGridSubsetPrediction
+from app.db.models import ModelRunGridSubsetPrediction
+from app.db.models.observations import HourlyActual
 
 
 def get_actuals_left_outer_join_with_predictions(*args):  # pylint: disable=unused-argument
